move ltw basic test case to package msalonly.ltw
